// src/base/error/types.ts
export enum ApiErrorCode {
  GLO_000 = 'GLO_000',
  GLO_001 = 'GLO_001',
  GLO_002 = 'GLO_002',
  GLO_003 = 'GLO_003',
  GLO_004 = 'GLO_004',
  GLO_005 = 'GLO_005',
  GLO_006 = 'GLO_006',
  GLO_007 = 'GLO_007',
  GLO_008 = 'GLO_008',
  GLO_009 = 'GLO_009',
  GLO_010 = 'GLO_010',
  GLO_011 = 'GLO_011',
  GLO_012 = 'GLO_012',
  GLO_013 = 'GLO_013',
  GLO_014 = 'GLO_014',
  GLO_015 = 'GLO_015',
  MUT_001 = 'MUT_001',
  MUT_002 = 'MUT_002',
  ACC_001 = 'ACC_001',
  ACC_002 = 'ACC_002',
  ACC_003 = 'ACC_003',
  ACC_004 = 'ACC_004',
  ACC_005 = 'ACC_005',
  ACC_006 = 'ACC_006',
  ACC_007 = 'ACC_007',
  ACC_008 = 'ACC_008',
  ACC_009 = 'ACC_009',
  ACC_010 = 'ACC_010',
  ACC_011 = 'ACC_011',
  ACC_012 = 'ACC_012',
  ACC_013 = 'ACC_013',
  BSA_001 = 'BSA_001',
  BSA_002 = 'BSA_002',
  BSA_003 = 'BSA_003',
  BSA_004 = 'BSA_004',
  BSA_005 = 'BSA_005',
  BSA_006 = 'BSA_006',
  BSA_007 = 'BSA_007',
  API_001 = 'API_001',
  API_002 = 'API_002',
  API_003 = 'API_003',
  API_004 = 'API_004',
  API_005 = 'API_005',
  REC_001 = 'REC_001',
  REC_002 = 'REC_002',
  REC_003 = 'REC_003',
  REC_004 = 'REC_004',
  REC_005 = 'REC_005',
  REC_006 = 'REC_006',
  REC_007 = 'REC_007',
  PAY_001 = 'PAY_001',
  PAY_002 = 'PAY_002',
  AUTH_001 = 'AUTH_001',
  AUTH_002 = 'AUTH_002',
  AUTH_003 = 'AUTH_003',
  IPN_001 = 'IPN_001',
  BSM_001 = 'BSM_001',
  BSM_002 = 'BSM_002',
  BSM_003 = 'BSM_003',
  BSM_004 = 'BSM_004',
  BSM_005 = 'BSM_005',
  BSM_006 = 'BSM_006',
  BSM_007 = 'BSM_007',
  BSM_008 = 'BSM_008',
  BSM_009 = 'BSM_009',
  BSS_001 = 'BSS_001',
  BSS_002 = 'BSS_002',
  BSS_003 = 'BSS_003',
  BSS_004 = 'BSS_004',
  BSS_005 = 'BSS_005',
  BSS_006 = 'BSS_006',
  BSS_007 = 'BSS_007',
  BSS_008 = 'BSS_008',
  BSS_009 = 'BSS_009',
  BSS_010 = 'BSS_010',
  BSS_011 = 'BSS_011',
  BSS_012 = 'BSS_012',
  BSS_013 = 'BSS_013',
  BSS_014 = 'BSS_014',
  BSB_001 = 'BSB_001',
  BSB_002 = 'BSB_002',
  BSW_001 = 'BSW_001',
  BSW_002 = 'BSW_002',
  BSW_003 = 'BSW_003',
  BSW_004 = 'BSW_004',
  BSW_005 = 'BSW_005',
  BSW_006 = 'BSW_006',
  BSW_007 = 'BSW_007',
  BSW_008 = 'BSW_008',
  BSW_009 = 'BSW_009',
  BSW_010 = 'BSW_010',
  BSW_011 = 'BSW_011',
  BSW_012 = 'BSW_012',
  BSGC_001 = 'BSGC_001',
  BSGC_002 = 'BSGC_002',
  BSGC_003 = 'BSGC_003',
  BAFF_001 = 'BAFF_001',
  BAFF_002 = 'BAFF_002',
  BAFF_003 = 'BAFF_003',
  BAFF_004 = 'BAFF_004',
  BSSOC_001 = 'BSSOC_001',
  BSSUP_001 = 'BSSUP_001',
  BSSUP_002 = 'BSSUP_002',
  REQ_000 = 'REQ_000',
}

export enum ApiErrorPath {
  GLO_000 = 'global.internal',
  GLO_001 = 'global.not_allowed',
  GLO_002 = 'global.access_denied',
  GLO_003 = 'global.validator_error',
  GLO_004 = 'global.wrong_body',
  GLO_005 = 'global.wrong_token',
  GLO_006 = 'global.missing_token',
  GLO_007 = 'global.access_denied_acl',
  GLO_008 = 'global.connection_timeout',
  GLO_009 = 'global.login_required',
  GLO_010 = 'global.not_found',
  GLO_011 = 'global.wrong_value',
  GLO_012 = 'global.token_expired',
  GLO_013 = 'global.service_not_available',
  GLO_014 = 'global.twofa_required',
  GLO_015 = 'global.missing_login_token',
  MUT_001 = 'mutex.long_lock',
  MUT_002 = 'mutex.exclusive_lock',
  ACC_001 = 'account.banned',
  ACC_002 = 'account.blocked',
  ACC_003 = 'account.wrong_login_or_password',
  ACC_004 = 'account.wrong_login',
  ACC_005 = 'account.wrong_password',
  ACC_006 = 'account.missing_email',
  ACC_007 = 'account.same_password_used',
  ACC_008 = 'account.wrong_twofa',
  ACC_009 = 'account.twofa_not_created',
  ACC_010 = 'account.twofa_already_created',
  ACC_011 = 'account.twofa_limit',
  ACC_012 = 'account.not_exists',
  ACC_013 = 'account.migration',
  BSA_001 = 'account.apikey_not_created',
  BSA_002 = 'account.apikey_already_created',
  BSA_003 = 'account.deletion_already_started',
  BSA_004 = 'account.external_withdraw_not_available',
  BSA_005 = 'account.kyc_wrong_step',
  BSA_006 = 'account.kyc_required',
  BSA_007 = 'account.action_not_allowed',
  API_001 = 'api.missing_apikey',
  API_002 = 'api.missing_secret',
  API_003 = 'api.wrong_ip_address',
  API_004 = 'api.private_ip_address',
  API_005 = 'api.wrong_apikey',
  REC_001 = 'record.item_not_found',
  REC_002 = 'record.item_not_owned',
  REC_003 = 'record.own_item',
  REC_004 = 'record.already_exists',
  REC_005 = 'record.not_exists',
  REC_006 = 'record.not_updated',
  REC_007 = 'record.forbidden',
  PAY_001 = 'payment.card_scoring',
  PAY_002 = 'payment.payment_issue',
  AUTH_001 = 'auth.failed_to_authenticate',
  AUTH_002 = 'auth.not_valid_identity',
  AUTH_003 = 'auth.authentication_failed',
  IPN_001 = 'ipn.verification_issue',
  BSM_001 = 'market.wrong_item_status',
  BSM_002 = 'market.item_already_queued',
  BSM_003 = 'market.max_price_limit',
  BSM_004 = 'market.item_not_available',
  BSM_005 = 'market.plan_not_found',
  BSM_006 = 'market.no_items_found',
  BSM_007 = 'market.items_limit',
  BSM_008 = 'market.instant_sell_disabled',
  BSM_009 = 'market.price_too_low',
  BSS_001 = 'steam.wrong_tradelink',
  BSS_002 = 'steam.private_inventory',
  BSS_003 = 'steam.cannot_check_tradeban',
  BSS_004 = 'steam.cannot_check_escrow',
  BSS_005 = 'steam.vac_ban',
  BSS_006 = 'steam.community_ban',
  BSS_007 = 'steam.economy_ban',
  BSS_008 = 'steam.private_profile',
  BSS_009 = 'steam.escrow',
  BSS_010 = 'steam.wrong_token',
  BSS_011 = 'steam.missing_token',
  BSS_012 = 'steam.missing_steam_id',
  BSS_013 = 'steam.platform_offline',
  BSS_014 = 'steam.fresh_account',
  BSB_001 = 'bot.not_found',
  BSB_002 = 'bot.not_available',
  BSW_001 = 'wallet.cannot_transfer_money',
  BSW_002 = 'wallet.amount_negative',
  BSW_003 = 'wallet.amount_positive',
  BSW_004 = 'wallet.card_scoring_too_low',
  BSW_005 = 'wallet.not_enough_balance',
  BSW_006 = 'wallet.kyc_window_limit',
  BSW_007 = 'wallet.service_window_limit',
  BSW_008 = 'wallet.amount_limit',
  BSW_009 = 'wallet.wrong_wallet_address',
  BSW_010 = 'wallet.card_not_found',
  BSW_011 = 'wallet.cannot_process_transaction',
  BSW_012 = 'wallet.only_past_months_accepted',
  BSGC_001 = 'gift_code.not_found',
  BSGC_002 = 'gift_code.already_used',
  BSGC_003 = 'gift_code.own_code',
  BAFF_001 = 'affiliate.same_code',
  BAFF_002 = 'affiliate.code_taken',
  BAFF_003 = 'affiliate.not_activated',
  BAFF_004 = 'affiliate.not_enough_balance',
  BSSOC_001 = 'social.cannot_create_chat',
  BSSUP_001 = 'support.wrong_status',
  BSSUP_002 = 'support.ticket_limit',
  REQ_000 = 'request.failed',
}

export enum ApiErrorMessage {
  GLO_000 = 'Internal error',
  GLO_001 = 'Action not allowed',
  GLO_002 = 'Access denied',
  GLO_003 = 'Validator error',
  GLO_004 = 'Wrong body',
  GLO_005 = 'Wrong token',
  GLO_006 = 'Missing token',
  GLO_007 = 'Access denied (ACL)',
  GLO_008 = 'Connection timeout',
  GLO_009 = 'Login required',
  GLO_010 = 'Not found',
  GLO_011 = 'Wrong value',
  GLO_012 = 'Token expired',
  GLO_013 = 'Service not available',
  GLO_014 = '2FA authentication required',
  GLO_015 = 'Login required',
  MUT_001 = 'Long lock, try again in seconds',
  MUT_002 = 'Exclusive lock',
  ACC_001 = 'Account banned',
  ACC_002 = 'Account blocked',
  ACC_003 = 'Wrong login or password',
  ACC_004 = 'Wrong login',
  ACC_005 = 'Wrong password',
  ACC_006 = 'Email address not assigned',
  ACC_007 = 'The same password was used',
  ACC_008 = 'Wrong 2FA code',
  ACC_009 = '2FA not created',
  ACC_010 = '2FA already created',
  ACC_011 = '2FA limit',
  ACC_012 = 'Not exists',
  ACC_013 = 'Migration',
  BSA_001 = 'API Key not created',
  BSA_002 = 'API Key already created',
  BSA_003 = 'Account deletion already started',
  BSA_004 = 'External withdraw is not available',
  BSA_005 = 'KYC wrong step',
  BSA_006 = 'KYC required',
  BSA_007 = 'Action not allowed',
  API_001 = 'Missing API Key',
  API_002 = 'Missing API Secret',
  API_003 = 'Wrong IP address',
  API_004 = 'Private IP address',
  API_005 = 'Wrong apikey',
  REC_001 = 'Item not found',
  REC_002 = 'Item not owned',
  REC_003 = 'Own item',
  REC_004 = 'Already exists',
  REC_005 = 'Not exists',
  REC_006 = 'Not updated',
  REC_007 = 'Forbidden',
  PAY_001 = 'Card scoring is too low',
  PAY_002 = 'Payment issue',
  AUTH_001 = 'Failed to authenticate',
  AUTH_002 = 'Not valid identity',
  AUTH_003 = 'Authentication failed',
  IPN_001 = 'Verification issue',
  BSM_001 = 'Wrong item status',
  BSM_002 = 'Item already queued',
  BSM_003 = 'Max price limit',
  BSM_004 = 'Item not available',
  BSM_005 = 'Plan not found',
  BSM_006 = 'No items found',
  BSM_007 = 'Items limit',
  BSM_008 = 'Instant sell disabled',
  BSM_009 = 'Price too low',
  BSS_001 = 'Wrong trade link',
  BSS_002 = 'Private Steam inventory',
  BSS_003 = 'We cannot check trade ban',
  BSS_004 = 'Cannot check escrow',
  BSS_005 = 'VAC ban',
  BSS_006 = 'Community ban',
  BSS_007 = 'Economy ban',
  BSS_008 = 'Private Steam profile',
  BSS_009 = 'Your account has escrow limit',
  BSS_010 = 'Wrong token',
  BSS_011 = 'Missing Steam trade url',
  BSS_012 = 'Missing Steam ID',
  BSS_013 = 'Platform offline',
  BSS_014 = 'You need to setup steam profile first',
  BSB_001 = 'Bot not found',
  BSB_002 = 'Bot not available',
  BSW_001 = 'Cannot transfer money',
  BSW_002 = 'Amount negative',
  BSW_003 = 'Amount positive',
  BSW_004 = 'Card scoring too low',
  BSW_005 = 'Not enough balance',
  BSW_006 = 'KYC window limit',
  BSW_007 = 'Service window limit',
  BSW_008 = 'Amount limit',
  BSW_009 = 'Wrong wallet address',
  BSW_010 = 'Card not found',
  BSW_011 = 'We cannot process your transaction',
  BSW_012 = 'Only past months accepted',
  BSGC_001 = 'Code not found',
  BSGC_002 = 'Code already used',
  BSGC_003 = 'This is your own code',
  BAFF_001 = 'Same code',
  BAFF_002 = 'This code is already taken',
  BAFF_003 = 'Not activated',
  BAFF_004 = 'Not enough balance',
  BSSOC_001 = 'Cannot create chat',
  BSSUP_001 = 'Wrong status',
  BSSUP_002 = 'Ticket limit',
  REQ_000 = 'Request failed',
}

export type IApiErrorType = 'public';

export interface IApiError {
  code: ApiErrorCode;
  error_type: IApiErrorType;
  error_path: ApiErrorPath;
  error_message: ApiErrorMessage;
}